Move and rename `currencyCodesToEdgeAssets`
